ready for build

+++ src/actions/deleteDataSource.ts
@@ -1,10 +1,16 @@
 'use server';
 
 import { prisma } from '@/lib/prisma';
+import { getServerSession } from 'next-auth';
 import { revalidatePath } from 'next/cache';
 
 export const deleteDataSource = async (id: number, prevState: any) => {
   try {
+    const session = await getServerSession();
+    if (!session) {
+      throw new Error('Kullanıcı bulunamadı.');
+    }
+
     await prisma.dataSource.delete({
       where: {
         id,
@@ -18,10 +24,18 @@ export const deleteDataSource = async (id: number, prevState: any) => {
     };
   } catch (e: any) {
     console.log(e);
+
+    const message = () => {
+      if (e?.message.includes('Foreign key constraint'))
+        return "Bu veri kaynağına bağlı dashboard'lar bulunuyor.";
+
+      return `Hata: ${e?.message}`;
+    };
+
     return {
       success: false,
       tried: true,
-      message: `Veri kaynağı silinemedi. Hata: ${e?.message}`,
+      message: `Veri kaynağı silinemedi. ${message()}`,
     };
   }
 };
